requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.4
google-api-core==2.11.1
google-api-python-client==2.92.0
google-auth==2.21.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.1
gTTS==2.3.2
httplib2==0.22.0
idna==3.4
oauthlib==3.2.2
protobuf==4.23.4
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAudio==0.2.13
pygame==2.5.0
pyparsing==3.1.0
pyttsx3==2.90
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
SpeechRecognition==3.10.0
uritemplate==4.1.1
urllib3==1.26.16
